implement single dip switch cycle time
Each side of the cycle can now have its time set (in minutes) via a single 8-bit dip switch. The 4 most significant bits are dedicated to the on time, and the 4 least significant bits are dedicated to the off time. Values are read in traditional binary fashion, which forces a maximum value of 15 minutes for either side of the cycle.
Clock on pin D4, load on pin D5, data in on pin 6.